modules/gbp-automation/requirements.txt
google-auth>=2.25.0
google-api-python-client>=2.120.0
click>=8.1.0
pydantic>=2.5.0
anthropic>=0.40.0
Pillow>=10.2.0
httpx>=0.27.0
pytest>=8.0.0
